feat(schema.prisma): prisma downloaded and database almost finished
I have added prisma to the project and made the connection with a local database named as
trainya_app. Added all the models and all their relations with each other. Comments in
schema.prisma: all the relations in the model, your cardinality and wich model will relation with
it. I specified the normalization models. Just remaining add all the attributes in all models.